refactor(runtime): extract scene-loading systems into scene_systems.rs
Moves ProjectManifest/ProjectSection/WindowSection, resolve_primitives,
resolve_physics_bodies, handle_scene_load, and resolve_physics_bodies_world
out of main.rs into a shared scene_systems module, so the upcoming headless
test runtime can reuse identical scene-load/physics-resolution behavior.
Pure refactor, no behavior change. Adds empty test_mode/test_protocol/
test_query module stubs for later tasks.
