+++ crates/bsengine-runtime/src/test_mode.rs
@@ -1,0 +1,3 @@
+pub fn run_test_mode(_project_dir: &str) {
+    todo!("implemented in a later task")
+}
